refactor: authNavigation을 유틸에서 훅으로 전환
